# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for مقرأة سَنا الآي Quran Center Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /calendar-events:
    get:
      operationId: listCalendarEvents
      tags: [calendar]
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Calendar events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
    post:
      operationId: createCalendarEvent
      tags: [calendar]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCalendarEventBody"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
  /calendar-events/{id}:
    patch:
      operationId: updateCalendarEvent
      tags: [calendar]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCalendarEventBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
    delete:
      operationId: deleteCalendarEvent
      tags: [calendar]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /badge-events:
    get:
      operationId: listBadgeEvents
      tags: [badges]
      responses:
        "200":
          description: Badge events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BadgeEvent"
    post:
      operationId: createBadgeEvent
      tags: [badges]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBadgeEventBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadgeEvent"
  /badge-events/{id}/auto-assign:
    post:
      operationId: autoAssignBadgeEvent
      tags: [badges]
      summary: Auto-assign badge to all qualifying entities
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Auto-assigned result
          content:
            application/json:
              schema:
                type: object
                properties:
                  assigned:
                    type: integer
                  skipped:
                    type: integer
                  message:
                    type: string
                required: [assigned, skipped, message]
  /badge-events/{id}:
    patch:
      operationId: updateBadgeEvent
      tags: [badges]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBadgeEventBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadgeEvent"
    delete:
      operationId: deleteBadgeEvent
      tags: [badges]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /badge-assignments:
    get:
      operationId: listBadgeAssignments
      tags: [badges]
      parameters:
        - name: entityType
          in: query
          schema:
            type: string
        - name: entityId
          in: query
          schema:
            type: integer
        - name: badgeEventId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Badge assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BadgeAssignment"
    post:
      operationId: createBadgeAssignment
      tags: [badges]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBadgeAssignmentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadgeAssignment"
  /badge-assignments/{id}:
    delete:
      operationId: deleteBadgeAssignment
      tags: [badges]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
...
  schemas:
...
    CalendarEvent:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        date:
          type: string
        endDate:
          type: ["string", "null"]
        color:
          type: string
        eventType:
          type: string
        description:
          type: ["string", "null"]
        createdById:
          type: integer
        createdAt:
          type: string
      required: [id, title, date, color, eventType, createdById, createdAt]
    CreateCalendarEventBody:
      type: object
      properties:
        title:
          type: string
        date:
          type: string
        endDate:
          type: ["string", "null"]
        color:
          type: string
        eventType:
          type: string
        description:
          type: ["string", "null"]
      required: [title, date, color, eventType]
...
    BadgeEvent:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        emoji:
          type: string
        color:
          type: string
        targetType:
          type: string
        dateFrom:
          type: string
        dateTo:
          type: string
        isActive:
          type: boolean
        createdById:
          type: integer
        createdAt:
          type: string
        assignmentCount:
          type: integer
      required: [id, name, emoji, color, targetType, dateFrom, dateTo, isActive, createdById, createdAt, assignmentCount]
    CreateBadgeEventBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        emoji:
          type: string
        color:
          type: string
        targetType:
          type: string
        dateFrom:
          type: string
        dateTo:
          type: string
        isActive:
          type: boolean
      required: [name, emoji, color, targetType, dateFrom, dateTo]
    BadgeAssignment:
      type: object
      properties:
        id:
          type: integer
        badgeEventId:
          type: integer
        badgeName:
          type: string
        badgeEmoji:
          type: string
        badgeColor:
          type: string
        entityType:
          type: string
        entityId:
          type: integer
        entityName:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, badgeEventId, badgeName, badgeEmoji, badgeColor, entityType, entityId, entityName, createdAt]
    CreateBadgeAssignmentBody:
      type: object
      properties:
        badgeEventId:
          type: integer
        entityType:
          type: string
        entityId:
          type: integer
        entityName:
          type: string
        notes:
          type: ["string", "null"]
      required: [badgeEventId, entityType, entityId, entityName]
...